+++ src/navigation/LibraryStack.js
@@ -1,5 +1,6 @@
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import LibraryScreen from '../screens/LibraryScreen';
+import BookEditScreen from '../screens/BookEditScreen';
 
 const Stack = createNativeStackNavigator();
 
@@ -9,6 +10,13 @@ export default function LibraryStack() {
       <Stack.Screen
         name="LibraryMain"
         component={LibraryScreen}
+        options={{
+          headerShown: false,
+        }}
+      />
+      <Stack.Screen
+        name="BookEdit"
+        component={BookEditScreen}
         options={{
           headerShown: false,
         }}